Struct procrs::pid::stat::PidStat

pub struct PidStat {
    pub pid: TaskId,
    pub comm: String,
    pub state: PidState,
    pub ppid: TaskId,
    pub pgrp: i32,
    pub session: i32,
    pub tty_nr: i32,
    pub tpgid: i32,
    pub flags: u32,
    pub minflt: u64,
    pub cminflt: u64,
    pub majflt: u64,
    pub cmajflt: u64,
    pub utime: u64,
    pub stime: u64,
    pub cutime: i64,
    pub cstime: i64,
    pub priority: i64,
    pub nice: i64,
    pub num_threads: i64,
    pub itrealvalue: i64,
    pub starttime: u64,
    pub vsize: u64,
    pub rss: i64,
    pub rsslim: u64,
    pub startcode: u64,
    pub endcode: u64,
    pub startstack: u64,
    pub kstkesp: u64,
    pub kstkeip: u64,
    pub signal: u64,
    pub blocked: u64,
    pub sigignore: u64,
    pub sigcatch: u64,
    pub wchan: u64,
    pub nswap: u64,
    pub cnswap: u64,
    pub exit_signal: Option<i32>,
    pub processor: Option<i32>,
    pub rt_priority: Option<u32>,
    pub policy: Option<u32>,
    pub delayacct_blkio_ticks: Option<u64>,
    pub guest_time: Option<u64>,
    pub cguest_time: Option<i64>,
    pub start_data: Option<u64>,
    pub end_data: Option<u64>,
    pub start_brk: Option<u64>,
    pub arg_start: Option<u64>,
    pub arg_end: Option<u64>,
    pub env_start: Option<u64>,
    pub env_end: Option<u64>,
    pub exit_code: Option<i32>,
}

A struct containing information from the stat file for a process.

This struct contains information from the /proc/[pid]/stat or /proc/[tgid]/task/[tid]/stat file, for a specific pid or tgid/tid. for a specific pid.

Fields

pid

The process id.

comm

The filename of the executable.

state

The process state.

ppid

The process id of the parent process.

pgrp

The process group id.

session

The session id of this process.

tty_nr

The controlling tty of this process.

tpgid

The id of the process controlling the tty of this process.

flags

The kernel flags of this processj.

minflt

Count of minor page faults not requiring disk access.

cminflt

Count of minor page faults in children we are waiting for.

majflt

Count of major page faults not requiring disk acceess.

cmajflt

Count of major page faults in children we are waiting for.

utime

Amout of time this process has been scheduled in user mode.

stime

Amount of time this process has been scheduled in kernel mode.

cutime

Amount of time children we are waiting for have been scheduled in user mode.

cstime

Amount of time children we are waiting for have been scheduled in kernel mode.

priority

Priority of process.

nice

Process nice value (19 low -> -20 high).

num_threads

Number of threads this process is using.

itrealvalue

Count of jiffies before we receive the next SIGALRM (0 since kernel 2.6.17).

starttime

The time the process started after boot (ticks since kernel 2.6).

vsize

Virtual memory size in bytes.

rss

Resident set size in pages.

rsslim

RSS soft limit of process.

startcode

Memory address where executable memory starts.

endcode

Memory address where executable memory stops.

startstack

The start address of the stack (ie: bottom).

kstkesp

The current value of ESP (stack pointer).

kstkeip

The current EIP (instruction pointer).

signal

The bitmap of pending signals, displayed as a decimal number.

blocked

The bitmap of blocked signals, displayed as a decimal number.

sigignore

The bitmap of ignored signals, displayed as a decimal number.

sigcatch

The bitmap of caught signals, displayed as a decimal number.

wchan

This is the "channel" in which the process is waiting.

nswap

Number of pages swapped (not maintained).

cnswap

Cumulative nswap for child processes (not maintained).

exit_signal

Signal sent to the parent when we die.

processor

CPU number last executed on.

rt_priority

Real-time scheduling priority.

policy

Scheduling policy.

delayacct_blkio_ticks

Aggregated block I/O delays, measured in clock ticks.

guest_time

Time spent running a virtual CPU for a guest OS, in clock ticks.

cguest_time

Guest time of the process's children, in clock ticks.

start_data

Address above which init'd and uninit'd (BSS) data is placed.

end_data

Address below which init'd and uninit'd (BSS) data is placed.

start_brk

Address above which program heap can be expanded.

arg_start

Address above which program cmdline args (arv) are placed.

arg_end

Address below which program cmdline args (arv) are placed.

env_start

Address above which program environment is placed.

env_end

Address below which program environment is placed.

exit_code

The thread's exit status.

Methods

impl PidStat

fn new(pid_dir: &Path) -> Result<Self, ProcError>

Generate PidStat struct given a process directory.